GGRC-540: Requests tests converted into Assessment
