+++ test/integration/ggrc/converters/test_csvs/assessment_full_no_warnings.csv
@@ -1,0 +1,48 @@
+Object type,,,,,,,,,,,,,,
+Assessment,Code*,Audit*,Title*,Description,Notes,State*,Assessor*,Creator*,Recipients,Send by default,url,evidence,map:project,map: Market
+,Assessment 1,Audit,Assessment title 1,Some desc 1,notes 1,NOT STARTED,"user1@a.com
+user2@a.com","user2@a.com
+user5@a.com","Verifier, Requester",Yes,http://i.imgur.com/Lppr247.jpg,"http://i.imgur.com/Lppr247.jpg evidence 啕 title 1
+http://i.imgur.com/qA0l5tb.gifv some title 2","proj-1
+proj-2
+proj-3",mkt-1
+,Assessment 2,Audit,Assessment title 2,Some desc 2,notes 2,In Progress,user1@a.com,user2@a.com,,No,,,proj-1,mkt-2
+,Assessment 3,Audit,Assessment title 3,Some desc 3,notes 3,Not StarTED,user1@a.com,user2@a.com,,,,,proj-2,mkt-3
+,Assessment 4,Audit,Assessment title 4,Some desc 4,notes 4,In PROGRESS,user1@a.com,user2@a.com,,,,,proj-3,
+,Assessment 5,Audit,Assessment title 5,,,not started,user1@a.com,user2@a.com,,,,,,
+,Assessment 6,Audit,Assessment title 6,,,not started,user1@a.com,user2@a.com,,,,,,
+,Assessment 7,Audit,Assessment title 7,,,not started,user1@a.com,user2@a.com,,,,,,
+,Assessment 8,Audit,Assessment title 8,,,not started,user1@a.com,user2@a.com,,,,,,
+,Assessment 9,Audit,Assessment title 9,,,not started,user1@a.com,user2@a.com,,,,,,
+,Assessment 10,Audit,Assessment title 10,,,In Progress,user1@a.com,user2@a.com,,,,,,
+,Assessment 11,Audit,Assessment title 11,,,In PROGRESS,user1@a.com,user2@a.com,,,,,,
+,Assessment 12,Audit,Assessment title 12,,,Not Started,user1@a.com,user2@a.com,,,,,,
+,,,,,,,,,,,,,,
+Object type,,,,,,,,,,,,,,
+Person,name,email,company,role,,,,,,,,,,
+,user 1,user1@a.com,google,Admin ,,,,,,,,,,
+,user 2,user2@a.com,google,gGRC Admin,,,,,,,,,,
+,user 3,user3@a.com,google,creator,,,,,,,,,,
+,user 4,user4@a.com,google,reader,,,,,,,,,,
+,user 5,user5@a.com,amazon,,,,,,,,,,,
+,,,,,,,,,,,,,,
+,,,,,,,,,,,,,,
+Object type,,,,,,,,,,,,,,
+Audit,Code*,Program*,Title*,Internal Audit Lead*,Status,,,,,,,,,
+,Audit,Program,Audit,user@example.com,Planned,,,,,,,,,
+,,,,,,,,,,,,,,
+Object type,,,,,,,,,,,,,,
+Program,Code*,Title*,Manager,,,,,,,,,,,
+,Program,Program,user@example.com,,,,,,,,,,,
+,,,,,,,,,,,,,,
+Object type,,,,,,,,,,,,,,
+Project,code*,title*,description,owner,state,,,,,,,,,
+,proj-1,proj-1,Project for Assessment testing 1,user1@a.com,Draft,,,,,,,,,
+,proj-2,proj-2,Project for Assessment testing 2,user2@a.com,Final,,,,,,,,,
+,proj-3,proj-3,Project for Assessment testing 3,user3@a.com,Effective,,,,,,,,,
+,,,,,,,,,,,,,,
+Object type,,,,,,,,,,,,,,
+Market,code*,title*,description,owner,state,,,,,,,,,
+,mkt-1,mkt-1,Market for Assessment testing 1,user3@a.com,Draft,,,,,,,,,
+,mkt-2,mkt-2,Market for Assessment testing 2,user4@a.com,Final,,,,,,,,,
+,mkt-3,mkt-3,Market for Assessment testing 3,user5@a.com,Effective,,,,,,,,,
