+++ test/integration/ggrc/converters/test_csvs/assessment_update_intermediate.csv
@@ -1,0 +1,17 @@
+Object type,,,,,,,,,,,,
+Assessment,Code*,Audit*,Title*,Description,Notes,State*,Assessor*,Creator*,map:project,map: Market,url,evidence
+,Assessment 1,Audit,Assessment title 1,Some desc 1,Edited notes,NOT STARTED,"user1@a.com
+user2@a.com","user2@a.com
+user3@a.com","proj-1
+proj-2
+proj-3",mkt-1,"a.b.com
+c.d.com",http://i.imgur.com/Lppr247.jpg evidence title 1
+,Assessment 2,Audit,Assessment title 2,Some desc 2,notes 2,Ready for Review,user1@a.com,user2@a.com,proj-1,mkt-2,,
+,Assessment 3,Audit,Assessment title 3,Some desc 3,notes 3,Completed,user1@a.com,user2@a.com,proj-2,mkt-3,,
+,Assessment 4,Audit,Assessment title 4,Some desc 4,notes 4,Verified,user1@a.com,user2@a.com,proj-3,,,
+,Assessment 5,Audit,Assessment title 1,,Edited notes,Ready for Review,user1@a.com,user2@a.com,,,,
+,Assessment 60,Audit,Assessment title 2,,,Not Started,user1@a.com,user2@a.com,,,,
+,Assessment 61,Audit,Assessment title 3,,,In Progress,user1@a.com,user2@a.com,,,,
+,Assessment 62,Audit,Assessment title 4,,,READY FOR REVIEW,user1@a.com,user2@a.com,,,,
+,Assessment 63,Audit,Assessment title 5,,,Verified,user1@a.com,user2@a.com,,,,
+,Assessment 64,Audit,Assessment title 6,,,COMPLETED,user1@a.com,user2@a.com,,,,
