+++ test/integration/ggrc/converters/test_csvs/assessment_with_warnings_and_errors.csv
@@ -1,0 +1,25 @@
+Object type,"This needs to be imported after assessment_full_no_warnings
+",,,These are the expected warnings,,,,,,,,,,
+Assessment,Code*,Audit*,Title*,Error description - non existing column will be ignored,Actual Error message,Notes,State*,Assessor*,Creator*,Finished Date,Verified Date,map:project,map: Market,url
+,Assessment 2,Audit,Assessment title 2,"you are not verifier, you can not verify this assessment",,notes 2,In Progress,user1@a.com,user2@a.com,7/2/2015,5/12/2016,,,http://i.imgur.com/Lppr247.jpg some invalid title
+,Assessment 3,Audit,Assessment title 3,"this assessment has changed and can not be verified in the first step, verification will be done only upon re-importing this csv", ,changed notes,In Progress,user1@a.com,user2@a.com,7/3/2015,5/13/2016,,,
+,Assessment 4,Audit,Assessment title 4,This should work and the state will be changed to verified,,notes 4,In Progress,user1@a.com,user2@a.com,7/4/2015,5/14/2016,,,
+,Assessment 5,Audit,Assessment title 1,State can not be verified if there is no verifier.,Could not change assessment state to verified.  There is no designated verifier for the assessment.,,In Progress,user1@a.com,user2@a.com,7/5/2015,5/15/2016,,,
+,Assessment 6,Audit,Assessment title 2,State can not be set to final if there is a verifier.,"Could not change assessment state to final.  When there are designated verifiers, the valid end state can only be 'verified'. ",,In Progress,user1@a.com,user2@a.com,7/6/2015,5/16/2016,,,
+,Assessment 7,Audit,Assessment title 3,No verifier and state going from open to final is fine,,,In Progress,user1@a.com,user2@a.com,7/7/2015,5/17/2016,,,
+,Assessment 8,Audit,Assessment title 4,This works normally.,,,In Progress,user1@a.com,user2@a.com,7/8/2015,5/18/2016,,,
+,Assessment 9,Audit,Assessment title 5,Invalid state finished if there is no verifier - shows only on change,,,In Progress,user1@a.com,user2@a.com,7/9/2015,5/19/2016,,,
+,Assessment 10,Audit,Assessment title 6,"This works because the change is removing verifier, and state does not change",,,In Progress,user1@a.com,user2@a.com,7/10/2015,5/20/2016,,,
+,Assessment 11,Audit,Assessment title 7,State can not transition back to open,,,In Progress,user1@a.com,user2@a.com,7/11/2015,5/21/2016,,,
+,Assessment 12,Audit,Assessment title 8,"State will be changed to in-progress, because of the updates to assessment.",,changes,In Progress,user1@a.com,user2@a.com,7/12/2015,5/22/2016,,,
+,Assessment 16,Audit,Assessment title 1,"proj-55 does not exist non_existing@a.com does not exist ", --- already exists ---,notes 1,In Progress,"user1@a.com
+user2@a.com","user2@a.com
+user3@a.com",7/1/2015,5/11/2016,proj-55,mkt-1,
+,Assessment 17,Audit,Assessment title 2,Missing Due On date - will default to today + 7 days,,notes 2,In Progress,user1@a.com,user2@a.com,5/11/2016,,proj-1,mkt-2,
+,Assessment 18,Audit,Assessment title 3,Missing Starts On on date - will default to today,,notes 2,In Progress,user1@a.com,user2@a.com,,5/12/2016,proj-1,mkt-2,
+,Assessment 19,Audit,Assessment title 4,"Missing Starts On on and Due On dates, default today / today + 7 days",,notes 3,In Progress,user1@a.com,user2@a.com,,,,,
+,Assessment 20,Audit,Assessment title 5,Missing assessment type - will take default , --- already exists ---,notes 4,In Progress,user1@a.com,user2@a.com,7/4/2015,5/14/2016,,,
+,Assessment 21,not existing,Assessment title 6,Not existing audit does not exist, --- already exists ---,notes 5,In Progress,user1@a.com,user2@a.com,7/5/2015,5/15/2016,,,
+,Assessment 22,Audit,same title no warning,"Can not set a state of a new assessment to verified or finished, because adding a new one is treated the same as updating an old one, except all values are new",,notes 3,open,user1@a.com,user2@a.com,7/3/2015,5/13/2016,proj-2,mkt-3,
+,Assessment 23,Audit,same title no warning,"Can not set a state of a new assessment to verified or finished, because adding a new one is treated the same as updating an old one, except all values are new",,notes 4,Completed,user1@a.com,user2@a.com,7/4/2015,5/14/2016,proj-3,,
+,Assessment 22,Audit,Assessment title 1,Duplicate code line will be ignored, --- already exists ---,,Verified,user1@a.com,user2@a.com,7/5/2015,5/15/2016,,,
