--- test/integration/ggrc/converters/test_csvs/request_full_no_warnings.csv
@@ -1,50 +1,0 @@
-Object type,,,,,,,,,,,,,,,,,,,
-Request,Code*,Audit*,Title*,Description,Notes,Status*,Assignee*,Requester*,Verifier,Request Type*,Starts On*,Due On*,Test,Recipients,Send by default,url,evidence,map:project,map: Market
-,Request 1,Audit,Request title 1,Some desc 1,notes 1,NOT STARTED,"user1@a.com
-user2@a.com","user2@a.com
-user3@a.com","user3@a.com
-user4@a.com
-user5@a.com",docuMEntation,7/1/2015,5/11/2016,Test text 1,"Verifier, Requester",Yes,http://i.imgur.com/Lppr247.jpg,"http://i.imgur.com/Lppr247.jpg evidence 啕 title 1
-http://i.imgur.com/qA0l5tb.gifv some title 2","proj-1
-proj-2
-proj-3",mkt-1
-,Request 2,Audit,Request title 2,Some desc 2,notes 2,In Progress,user1@a.com,user2@a.com,user3@a.com,interviEW,7/2/2015,5/12/2016,Test text 2,,No,,,proj-1,mkt-2
-,Request 3,Audit,Request title 3,Some desc 3,notes 3,Not StarTED,user1@a.com,user2@a.com,user@example.com,Documentation,7/3/2015,5/13/2016,Test text 3,,,,,proj-2,mkt-3
-,Request 4,Audit,Request title 4,Some desc 4,notes 4,In PROGRESS,user1@a.com,user2@a.com,user@example.com,Interview,7/4/2015,5/14/2016,Test text 4,,,,,proj-3,
-,Request 5,Audit,Request title 1,,,not started,user1@a.com,user2@a.com,,Interview,7/5/2015,5/15/2016,,,,,,,
-,Request 6,Audit,Request title 2,,,not started,user1@a.com,user2@a.com,user@example.com,Interview,7/6/2015,5/16/2016,,,,,,,
-,Request 7,Audit,Request title 3,,,not started,user1@a.com,user2@a.com,,Interview,7/7/2015,5/17/2016,,,,,,,
-,Request 8,Audit,Request title 4,,,not started,user1@a.com,user2@a.com,user@example.com,Interview,7/8/2015,5/18/2016,,,,,,,
-,Request 9,Audit,Request title 5,,,not started,user1@a.com,user2@a.com,,Interview,7/9/2015,5/19/2016,,,,,,,
-,Request 10,Audit,Request title 6,,,In Progress,user1@a.com,user2@a.com,user@example.com,Interview,7/10/2015,5/20/2016,,,,,,,
-,Request 11,Audit,Request title 7,,,In PROGRESS,user1@a.com,user2@a.com,user@example.com,Interview,7/11/2015,5/21/2016,,,,,,,
-,Request 12,Audit,Request title 8,,,Not Started,user1@a.com,user2@a.com,user@example.com,Interview,7/12/2015,5/22/2016,,,,,,,
-,,,,,,,,,,,,,,,,,,,
-Object type,,,,,,,,,,,,,,,,,,,
-Person,name,email,company,role,,,,,,,,,,,,,,,
-,user 1,user1@a.com,google,Admin ,,,,,,,,,,,,,,,
-,user 2,user2@a.com,google,gGRC Admin,,,,,,,,,,,,,,,
-,user 3,user3@a.com,google,creator,,,,,,,,,,,,,,,
-,user 4,user4@a.com,google,reader,,,,,,,,,,,,,,,
-,user 5,user5@a.com,amazon,,,,,,,,,,,,,,,,
-,,,,,,,,,,,,,,,,,,,
-,,,,,,,,,,,,,,,,,,,
-Object type,,,,,,,,,,,,,,,,,,,
-Audit,Code*,Program*,Title*,Internal Audit Lead*,Status,,,,,,,,,,,,,,
-,Audit,Program,Audit,user@example.com,Planned,,,,,,,,,,,,,,
-,,,,,,,,,,,,,,,,,,,
-Object type,,,,,,,,,,,,,,,,,,,
-Program,Code*,Title*,Manager,,,,,,,,,,,,,,,,
-,Program,Program,user@example.com,,,,,,,,,,,,,,,,
-,,,,,,,,,,,,,,,,,,,
-Object type,,,,,,,,,,,,,,,,,,,
-Project,code*,title*,description,owner,state,,,,,,,,,,,,,,
-,proj-1,proj-1,Project for Request testing 1,user1@a.com,Draft,,,,,,,,,,,,,,
-,proj-2,proj-2,Project for Request testing 2,user2@a.com,Final,,,,,,,,,,,,,,
-,proj-3,proj-3,Project for Request testing 3,user3@a.com,Effective,,,,,,,,,,,,,,
-,,,,,,,,,,,,,,,,,,,
-Object type,,,,,,,,,,,,,,,,,,,
-Market,code*,title*,description,owner,state,,,,,,,,,,,,,,
-,mkt-1,mkt-1,Market for Request testing 1,user3@a.com,Draft,,,,,,,,,,,,,,
-,mkt-2,mkt-2,Market for Request testing 2,user4@a.com,Final,,,,,,,,,,,,,,
-,mkt-3,mkt-3,Market for Request testing 3,user5@a.com,Effective,,,,,,,,,,,,,,
