--- test/integration/ggrc/converters/test_csvs/request_update_intermediate.csv
@@ -1,19 +1,0 @@
-Object type,,,,,,,,,,,,,,,,,
-Request,Code*,Audit*,Title*,Description,Notes,Status*,Assignee*,Requester*,Verifier,Request Type*,Starts On*,Due On*,Test,map:project,map: Market,url,evidence
-,Request 1,Audit,Request title 1,Some desc 1,Edited notes,NOT STARTED,"user1@a.com
-user2@a.com","user2@a.com
-user3@a.com","user3@a.com
-user4@a.com
-user5@a.com",docuMEntation,7/1/2015,5/11/2016,Test text 1,"proj-1
-proj-2
-proj-3",mkt-1,"a.b.com
-c.d.com",http://i.imgur.com/Lppr247.jpg evidence title 1
-,Request 2,Audit,Request title 2,Some desc 2,notes 2,Ready for Review,user1@a.com,user2@a.com,user3@a.com,interviEW,7/2/2015,5/12/2016,Test text 2,proj-1,mkt-2,,
-,Request 3,Audit,Request title 3,Some desc 3,notes 3,Completed,user1@a.com,user2@a.com,user@example.com,Documentation,7/3/2015,5/13/2016,Test text 3,proj-2,mkt-3,,
-,Request 4,Audit,Request title 4,Some desc 4,notes 4,Verified,user1@a.com,user2@a.com,user@example.com,Interview,7/4/2015,5/14/2016,Test text 4,proj-3,,,
-,Request 5,Audit,Request title 1,,Edited notes,Ready for Review,user1@a.com,user2@a.com,,Interview,7/5/2015,5/15/2016,,,,,
-,Request 60,Audit,Request title 2,,,Not Started,user1@a.com,user2@a.com,,Interview,7/6/2015,5/16/2016,,,,,
-,Request 61,Audit,Request title 3,,,In Progress,user1@a.com,user2@a.com,,Interview,7/7/2015,5/17/2016,,,,,
-,Request 62,Audit,Request title 4,,,READY FOR REVIEW,user1@a.com,user2@a.com,,Interview,7/8/2015,5/18/2016,,,,,
-,Request 63,Audit,Request title 5,,,Verified,user1@a.com,user2@a.com,,Interview,7/9/2015,5/19/2016,,,,,
-,Request 64,Audit,Request title 6,,,COMPLETED,user1@a.com,user2@a.com,,Interview,7/10/2015,5/20/2016,,,,,
